requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
python-dotenv
aiohttp
geopy